add Video::NetSink functionality into Video::Input
